requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==12.0.0
scikit-image==0.25.2
scipy==1.15.3
tifffile==2025.5.10
Werkzeug==3.1.3
supabase==2.24.0
dotenv==0.9.9
psycopg2-binary